% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LaLonde}
\alias{LaLonde}
\title{LaLonde data set}
\format{A data frame with 3212 observations and 12 variables.
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
This version of the data is from \code{\link[CBPS]{LaLonde}} data in \code{CBPS} package.
}
\usage{
LaLonde
}
\description{
Data from the National Supported Work Demonstration Program and 
  the Panel Study for Income Dynamics. A benchmark data set, which is
  extensively analyzed by Lalonde (1986), Dehejia and Wahba (1999), and Imai
  and Ratkovic (2014).
}
\keyword{datasets}
