% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nawt.R
\name{nawt}
\alias{nawt}
\title{Navigated Weighting (NAWT) Estimation}
\usage{
nawt(formula, outcome, estimand = "ATT", method = "score", data,
  weights = NULL, alpha = 2, twostep = TRUE, boot = FALSE,
  B = 2000, clevel = 0.95, message = TRUE)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}} (or one that
can be coerced to that class): a symbolic description of the model to be 
fitted.}

\item{outcome}{a character string specifying the name of outcome values
in \code{data}.}

\item{estimand}{a character string specifying a parameter of interest. Choose
"ATT" for the average treatment effect on treated estimation, "ATE" for the
average treatment effect estimation, or "MO" for the average outcomes 
estimation in missing outcome cases. You can choose "ATEcombined" for the  
combined estimation for the average treatment effect estimation.}

\item{method}{a character string specifying a type of weignting functions in 
propensity score estimation (\eqn{\omega(\pi)}). Choose "score" for a power
function of propensity scores (need to specify the value for alpha), "cb"
for a covariate balancing weighting function, or "both" to use both the
above weighting functions (need to specify the value for alpha).}

\item{data}{a data frame containing the outcomes and the variables in the 
model.}

\item{weights}{an optional vector of ‘prior weights’ (e.g. sampleing weights)
to be used in the fitting process. Should be NULL or a numeric vector.}

\item{alpha}{a positive value for an exponent in a power weighting function.
Default is 2. Set to 0 to use the standard logistic regression for  
propensity score estimation. Note that \code{nawt} with \eqn{m / 2} where 
\eqn{0 \le m \le 10} runs substantially faster than with any other values.}

\item{twostep}{a logical value indicating whether to use a two step estimator
when \code{method = "both"}. Default is \code{TRUE}. Set to \code{FALSE} 
to use a continuously-updating GMM estimator, which is substantially
computationally intensive.}

\item{boot}{a logical value indicating whether to use a non-parametric 
bootstrapping method to estimate the variance-covariance matrix and 
confidence intervals for parameters. Default is \code{FALSE}. Set to 
\code{FALSE} to use a sandwich-type asymptotic covariance estimator.}

\item{B}{the number of bootstrap replicates. Default is 2,000.}

\item{clevel}{confidence level. Default is 0.95.}

\item{message}{a logical value indicating whether messages are shown or not.}
}
\value{
\code{nawt} returns an object of class inheriting from "nawt". 

The function summary (i.e., \code{\link{summary.nawt}}) can be used to obtain or print a 
  summary of the results.

An object of class "nawt" is a list containing the following components:

\item{est}{the estimates of the parameter of interest.}
\item{weights}{the estimated inverse probability weights.}
\item{ps}{the estimated propensity scores. A matrix of two sets of the 
  estimated propensity scores is returned when \code{estimand = "ATE"}.}
\item{coefficients}{a named vector of coefficients. A matrix of two sets of
  coefficients for two sets of propensity scores is returned when 
  \code{estimand = "ATE"}.}
\item{varcov}{the variance-covariance matrix of the coefficients and 
  parameter of interest.}
\item{converged}{logical. Was the algorithm judged to have converged?}
\item{naive_weights}{the estimated inverse probability weights with the 
  standard logistic regression for the propensity score estimation.}
\item{naive_coef}{a named vector of coefficients with the standard logistic 
  regression for the propensity score estimation.}
\item{KLd}{the Kullback-Leibler divergence from the uniform base weights to 
  the estimated inverse probability weights.}
\item{scratio}{an optimal ratio of the covariate balancing weighting function
  to the power weighting function in taking the weighted average weights for
  the weighted score conditions when \code{method = "both"} and \code{twostep = TRUE}. 
  A vector of length two for two propensity score estimation is returned when
  \code{estimand = "ATE"}.}
\item{estimand}{the parameter of interest specified.}
\item{method}{the method specified.}
\item{outcome}{the outcome vector.}
\item{alpha}{alpha specified.}
\item{formula}{formula specified.}
\item{names.x}{names of the explanatory variables in propensity score 
  estimation.}
\item{prior.weights}{the weights initially supplied, a vector of 1s if none
  were.}
\item{treat}{the treatment vector. The missingness vector when the missing
outcome cases.}
\item{data}{the data argument.}
\item{ci}{a matrix of the confidence intervals for the parameter of interest.}
\item{omega}{a vetor of weights for the weighted score conditions (\eqn{\omega}).
  A matrix of two sets of omega is returned when \code{estimand = "ATE"}.}
\item{effN_ps}{The effective sample size for the propensity score estimation.
  A vector of length two for two propensity score esimtion is returned when 
  \code{estimand = "ATE"}.}
\item{effN_est}{The effective sample size for the parameter of interest
  estimation.}
\item{effN_original}{The effective sample size with the initial weights.}
\item{call}{the matched call.}
}
\description{
\code{nawt} estimates a pre-specified parameter of interest (e.g., the 
average treatment effects or the average treatment effects on the treated) 
with the inverse probability weighting where propensity scores are estimated
using estimating equations suitable for the parameter of interest. It 
includes the covariate balancing propensity score proposed by Imai and 
Ratkovic (2014), which uses covariate balancing conditions in propensity 
score estimation. \code{nawt} can also be used to estimate average outcomes
in missing outcome cases.
}
\details{
DETAILS
}
