% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nawt.R
\name{print.nawt}
\alias{print.nawt}
\title{Print navigated weighting estimation}
\usage{
\method{print}{nawt}(x, ...)
}
\arguments{
\item{x}{an object of class “nawt”, usually, a result of a call to \code{\link{nawt}}.}

\item{...}{additional arguments to be passed to print.}
}
\value{
No retrun value, called for side effects.
}
\description{
Prints a fitted \code{nawt} object.
}
\seealso{
\code{\link{nawt}}, \code{\link[base]{print}}
}
\author{
Hiroto Katsumata
}
