% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.nawt.R
\name{summary.nawt}
\alias{summary.nawt}
\title{Summarizing Navigated Weighting Estimation}
\usage{
\method{summary}{nawt}(result)
}
\description{
Prints a summary of a fitted \code{nawt} object.
}
\details{
Prints a summmary of a \code{nawt} object, in a format similar to glm.
}
