% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nawt.R
\name{nawt}
\alias{nawt}
\title{Navigated Weighting (NAWT) Estimation}
\usage{
nawt(formula, outcome, estimand = "ATT", method = "score", data,
  weights = NULL, alpha = 2, twostep = TRUE, boot = FALSE,
  B = 2000, clevel = 0.95, message = TRUE)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}} (or one that
can be coerced to that class): a symbolic description of the model to be 
fitted.}

\item{outcome}{a character string specifying the name of outcome values
in \code{data}}

\item{estimand}{a character string specifying a parameter of interest. Choose
"ATT" for the average treatment effect on treated estimation, "ATE" for the
average treatment effect estimation, or "MO" for the average outcomes 
estimation in missing outcome cases. You can choose "ATEcombined" for the  
combined estimation for the average treatment effect estimation.}

\item{method}{a character string specifying a type of weignting functions in 
propensity score estimation (\eqn{\omega(\pi)}). Choose "score" for a power
function of propensity scores (need to specify the value for alpha), "cb"
for a covariate balancing weighting function, or "both" to use both the
above weighting functions (need to specify the value for alpha).}

\item{data}{a data frame containing the outcomes and the variables in the 
model.}

\item{weights}{an optional vector of ‘prior weights’ (e.g. sampleing weights)
to be used in the fitting process. Should be NULL or a numeric vector.}

\item{alpha}{a positive value for an exponent in a power weighting function.
Default is 2. Set to 0 to use the standard logistic regression for  
propensity score estimation. Note that \code{nawt} with \eqn{m / 2} where 
\eqn{0 \le m \le 10} runs substantially faster than with any other values.}

\item{twostep}{a logical value indicating whether to use a two step estimator
when \code{method = "both"}. Default is \code{TRUE}. Set to \code{FALSE} 
to use a continuously-updating GMM estimator, which is substantially
computationally intensive.}

\item{boot}{a logical value indicating whether to use a non-parametric 
bootstrapping method to estimate the variance-covariance matrix and 
confidence intervals for parameters. Default is \code{FALSE}. Set to 
\code{FALSE} to use a sandwich-type asymptotic covariance estimator.}

\item{B}{the number of bootstrap replicates. Default is 2,000.}

\item{clevel}{confidence level. Default is 0.95.}

\item{message}{a logical value indicating whether messages are shown or not.}
}
\description{
\code{nawt} estimates a pre-specified parameter of interest (e.g., the 
average treatment effects or the average treatment effects on the treated) 
with the inverse probability weighting where propensity scores are estimated
using estimating equations suitable for the parameter of interest. It 
includes the covariate balancing propensity score proposed by Imai and 
Ratkovic (2014), which uses covariate balancing conditions in propensity 
score estimation. \code{nawt} can also be used to estimate average outcomes
in missing outcome cases.
}
\details{
DETAILS
}
