% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.nawt.R
\name{summary.nawt}
\alias{summary.nawt}
\title{Summarizing navigated weighting estimation}
\usage{
\method{summary}{nawt}(object, ...)
}
\arguments{
\item{object}{an object of class “nawt”, usually, a result of a call to \code{\link{nawt}}.}

\item{...}{additional arguments to be passed to summary.}
}
\value{
\item{call}{the matched call.}
\item{est}{the point estimate of the parameter of interest.}
\item{coefficients}{a table including coefficients, standard errors, 
  z-values, and two-sided p-values.}
\item{effN_ps}{the effective sample size for the propensity score estimation.}
\item{effN_est}{the effective sample size for the parameter of interest
  estimation.}
}
\description{
Prints a summary of a fitted \code{nawt} object.
}
\details{
Prints a summary of a \code{nawt} object, in a format similar to glm.
}
\examples{
# For examples see example(nawt)
}
\seealso{
\code{\link{nawt}}, \code{\link[base]{summary}}
}
\author{
Hiroto Katsumata
}
